+++ back-end/service/player.service.ts
@@ -2,15 +2,18 @@ import playerDb from "../repository/player.db"
 import { Player } from "../model/player";
 import { PlayerInput } from "../types/types";
 
-const getAllPlayers = async (): Promise<Player[]> => {
-    // if (!email) {
-    //     throw new Error('Cooked token not found');
-    // }
-
+const getAllPlayers = async ({email}: {email: string}): Promise<Player[]> => {
+    if (!email) {
+        throw new Error('Cooked token not found');
+    }
     return playerDb.findAll();
 }
 
-const getPlayerById = async (id: number): Promise<Player> => {
+const getPlayerById = async (id: number, {email} : {email: string}): Promise<Player> => {
+    if (!email) {
+        throw new Error('Cooked token not found');
+    }
+
     const player = await playerDb.findById(id);
     if (!player) {
         throw new Error(`Player with id ${id} not found`);
@@ -18,19 +21,34 @@ const getPlayerById = async (id: number): Promise<Player> => {
     return player;
 }
 
-const findPlayerByNumber = async (number: number): Promise<Player | undefined> => {
-    return playerDb.findByNumber(number);
-}
 
-const addPlayer = async ({name, number ,position, birthdate, imageUrl, teamId, stat}: PlayerInput): Promise<Player> => {
-    if (await findPlayerByNumber(number)) {
+
+const addPlayer = async ({name, number ,position, birthdate, imageUrl, teamId, stat}: PlayerInput, {email, role}: {email :string, role: string}): Promise<Player> => {
+    if (!email) {
+        throw new Error('Cooked token not found');
+    }
+
+    if (role == 'Player') {
+        throw new Error('You do not have the permission to add a player');
+    }
+
+    if (await playerDb.findById(number)) {
         throw new Error(`Player with number ${number} already exists`);
     }
 
     return playerDb.addPlayer({name, number, position, birthdate, imageUrl, teamId, stat});
 }
 
-const updatePlayer = async (id: number, {name, number, position, birthdate}: PlayerInput): Promise<Player> => {
+const updatePlayer = async (id: number, {name, number, position, birthdate}: PlayerInput, {email, role}: {email :string, role: string} ): Promise<Player> => {
+
+     if (!email) {
+        throw new Error('Cooked token not found');
+    }
+
+    if (role == 'Player') {
+        throw new Error('You do not have the permission to add a player');
+    }
+
     if (await playerDb.findByNumber(number)) {
         throw new Error(`Player with number ${number} already exists`);
     }
@@ -38,7 +56,15 @@ const updatePlayer = async (id: number, {name, number, position, birthdate}: Pla
     return playerDb.updatePlayer(id, {name, number, position, birthdate,});
 }
 
-const RemovePlayer = async (id: number): Promise<void> => {
+const RemovePlayer = async (id: number, {email, role}: {email :string, role: string}): Promise<void> => {
+
+    if (!email) {
+        throw new Error('Cooked token not found');
+    }
+
+    if (role !== 'Admin') {
+        throw new Error('You do not have the permission to delete a player.');
+    }
    playerDb.deletePlayer(id);
 }
